+++ tests/test_category_detection.py
@@ -10,6 +10,10 @@ class DummyProvider(BaseSearchProvider):
     def id(self) -> str:
         return "dummy"
 
+    def indexers(self) -> list[tuple[str, str]]:
+        """Return dummy indexer."""
+        return [(self.id(), self.short_name)]
+
     @property
     def short_name(self) -> str:
         return "Dummy"